Update go.py
In this updated version, a separate random number generation is used for VC and Growth deals to ensure that changes in parameters of VC deals will not affect Growth deals and vice versa. Additionally, KDE (Kernel Density Estimate) curves have been added to the histogram to display the probability density functions of the return on investment distributions.