Update go.py
The monte_carlo_simulation function now takes an additional parameter, growth_skewness, and uses skewnorm.rvs() from SciPy to sample from a skew normal distribution for growth investments.

In the sidebar controls under "Growth Investments", a new slider control for "Growth Distribution Skewness" is added, allowing users to control the skewness of the growth deal returns distribution.

In the histogram plot, Kernel Density Estimate (KDE) curves are added for both VC Deals and Growth Deals by setting the kde=True parameter in the sns.histplot() function. This adds a smooth curve to the histogram which estimates the probability density function of the data.