Update go.py
Inside the monte_carlo_simulation function, I used the skewnorm distribution from the scipy.stats module to create a positively skewed normal distribution for growth investments. The skewnorm takes an additional parameter called a, which represents the skewness. Positive values represent positive skewness.
In the sidebar controls, I added an additional slider to control the skewness of the growth investment distribution. The slider allows for values between -10 and 10. Positive values result in positive skewness.
For the histogram, the kde=True parameter has been added to sns.histplot calls to display the probability density function curve along with the histogram bars.