Update go.py
It seems that there was a problem with the scaling of the skewed normal distribution for the growth investments. I have updated the code to better control the distribution of the growth investments based on the settings from the sidebar. Specifically, I updated the section where the scaled_value is computed from the skewed normal distribution.