Update go.py
 ValueError: zero-size array to reduction operation minimum which has no identity, occurs when the input array for the min() or max() function is empty.

In the generate_growth_deals() function, we're trying to normalize returns. However, if growth_deals is zero (which can happen in our loop where we simulate scenarios from all VC deals to all growth deals), no returns are generated and returns will be an empty array.

We should check if the returns array is empty before attempting to normalize the returns.