Update go.py
This code now utilizes a Zipf's distribution for the VC investments, which is a common way to model power-law distributions in empirical data. Note that the Zipf's parameter s governs the shape of the distribution. The code also continues to use the skewed normal distribution for the growth investments. All inputs are provided through the Streamlit interface.

Last code was problematic at vc returns