Update go.py
With this modification, the growth investments are modeled using a lognormal distribution. The sidebar controls Growth Log-Normal Mean (μ of log) and Growth Log-Normal Std Dev (σ of log) allow you to control the parameters of the lognormal distribution, and you will be able to observe a positively skewed distribution for the growth investments.