Update go.py
updated code with the addition of a new scatterplot chart displaying the relationship between the 75th percentile of the ROI and the Sharpe ratio for each number of growth deals:

In this updated code, a new scatterplot (fig6) is added after all the previous charts. This scatterplot shows the top quartile (75th percentile) of the ROI on the x-axis and the corresponding Sharpe ratios on the y-axis. This new chart helps to visualize how the Sharpe ratio varies with the top quartile of the ROI for different numbers of growth deals.

Last version broke down