Update go.py
Fixed mode error: 

The error message IndexError: invalid index to scalar variable. suggests that the mode() function is not returning an array-like structure that can be indexed with [0][0].

The mode() function from the SciPy library returns a ModeResult object that contains two arrays - mode which contains the modes and count which contains the count of each mode.